Output .d.ts next to js file for better editor support
