+++ tsconfig.esm.json
@@ -3,6 +3,7 @@
     "compilerOptions": {
         "target": "ES2019",
         "module": "ES2020",
-        "outDir": "./dist/esm"
+        "outDir": "./dist/esm",
+        "declaration": true,
     }
 }
