--- tsconfig.types.json
@@ -1,9 +1,0 @@
-{
-    "extends": "./tsconfig.base.json",
-    "compilerOptions": {
-        "target": "ESNext",
-        "outDir": "./dist/types",
-        "declaration": true,
-        "emitDeclarationOnly": true
-    }
-}
